AI-helm/requirements.txt
numpy==1.21.0
tflite-runtime==2.7.0
scikit-learn==0.24.2
pandas==1.3.0
paho-mqtt==1.6.1
flask==3.1.0
requests==2.32.3